# Rework parsing and improve parse performance