--- a/cover.ps1
+++ b/cover.ps1
@@ -4,10 +4,10 @@ nuget install -OutputDirectory packages -Version 3.1.2 ReportGenerator
 
 $dotnetPath=(Join-Path $env:ProgramFiles (Join-Path dotnet dotnet.exe))
 
-& $dotnetPath build /p:RunCoverage=True
+& $dotnetPath build /p:RunCoverage=True -c:Debug
 
 .\packages\OpenCover.4.6.519\tools\OpenCover.Console.exe -register:user -target:"$dotnetPath" `
-    -targetargs:"test --no-build" -output:"coverage.xml" -oldstyle `
+    -targetargs:"test --no-build -c:Debug" -output:"coverage.xml" -oldstyle `
     -filter:"+[Ward*]* -[Ward.DnsClient]Asn1.* -[Ward.*.Tests]*"
 
 .\packages\ReportGenerator.3.1.2\tools\ReportGenerator.exe -reports:coverage.xml -targetdir:coverage