--- a/src/Ward.Console/Program.cs
+++ b/src/Ward.Console/Program.cs
@@ -18,7 +18,12 @@ namespace Ward.Console
                 return 1;
             }
 
-            var client = new UdpDnsClient("172.16.128.1", 53);
+            string dnsServer = "172.16.128.1";
+            if (args.Length == 3 && args[2][0] == '@')
+                dnsServer = args[2].Substring(1);
+
+            SConsole.WriteLine($"Querying {dnsServer} for {args[1]} records for {args[0]}.");
+            var client = new UdpDnsClient(dnsServer, 53);
             var resolve = await client.ResolveAsync(args[0], Enum.Parse<Dns.Type>(args[1]), Class.Internet);
 
             foreach (var record in resolve.Results)