--- a/src/Ward.Dns/Records/CaaRecord.cs
+++ b/src/Ward.Dns/Records/CaaRecord.cs
@@ -29,6 +29,7 @@ namespace Ward.Dns.Records
             Value = Encoding.ASCII.GetString((byte*)dataPointer + 2 + tagLength, (length - 2 - tagLength));
         }
 
+        [System.Diagnostics.DebuggerStepThrough]
         public override string ToString() =>
             $"{Name}\t{TimeToLive}\t{Class}\t{Type}\t{(Critical ? 1 : 0)} {Tag} {Value}";
     }