… .appveyor.yml
… benchmarks/Ward.Benchmarks/Config.cs
… benchmarks/Ward.Benchmarks/MessageParseBenchmark.cs
… src/Ward.Dns/Records/CnameRecord.cs
… src/Ward.Dns/Records/GenericRecord.cs
… src/Ward.Dns/Records/MailExchangerRecord.cs
… src/Ward.Dns/Records/NsRecord.cs
… src/Ward.Dns/Records/PtrRecord.cs
… src/Ward.Dns/Records/RecordFactory.cs
… src/Ward.Dns/Records/SoaRecord.cs
… src/Ward.Dns/Records/TxtRecord.cs
… src/Ward.Dns/ReturnCode.cs
… src/Ward.Dns/Utils.cs
… src/Ward.Dns/Ward.Dns.csproj
… src/Ward.DnsClient/HttpsDnsClient.cs
… src/Ward.DnsClient/IResolveResult.cs
… src/Ward.DnsClient/ResolveResult.cs
… src/Ward.DnsClient/TcpDnsClient.cs
… src/Ward.DnsClient/UdpDnsClient.cs
… src/Ward.DnsClient/Utils.cs
… test/Ward.Dns.Tests/Assertions.cs
… test/Ward.Dns.Tests/MessageTests.cs
… test/Ward.Dns.Tests/TestGenerators.cs
… test/Ward.DnsClient.Tests/HttpsClientTests.cs
… test/Ward.DnsClient.Tests/TcpClientTests.cs
… test/Ward.DnsClient.Tests/UtilsTests.cs
… test/Ward.Tests.Core/TestCase.cs
… test/Ward.Tests.Core/TestCaseData.toml
… test/Ward.Tests.Core/TestCaseLoader.cs