Reworks parsing to not be stream-based and to allocate as little as possible,
with an eye toward improved parse performance. Initial tests suggest an improvement of roughly 30% on
.NET Core, and more on Mono/.NET Desktop.

Last `master` commit before this branch:

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.14393.2007 (1607/AnniversaryUpdate/Redstone1), VM=Hyper-V
Intel Xeon CPU E5-2697 v3 2.60GHz, 1 CPU, 2 logical and 2 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0
  Job-YTNNJT : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
  Job-RRKVEU : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0


```
|       Method |     Toolchain |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------- |-------------- |---------:|----------:|----------:|-------:|----------:|
| ParseMessage | .NET Core 2.0 | 2.219 us | 0.0424 us | 0.0376 us | 0.3815 |   2.36 KB |
| ParseMessage |  CsProjnet461 | 2.158 us | 0.0418 us | 0.0529 us | 0.3853 |   2.38 KB |

Tip of this branch:

``` ini

BenchmarkDotNet=v0.10.14, OS=Windows 10.0.14393.2007 (1607/AnniversaryUpdate/Redstone1), VM=Hyper-V
Intel Xeon CPU E5-2697 v3 2.60GHz, 1 CPU, 2 logical and 2 physical cores
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0
  Job-YTNNJT : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
  Job-RRKVEU : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0


```
|       Method |     Toolchain |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------- |-------------- |---------:|----------:|----------:|-------:|----------:|
| ParseMessage | .NET Core 2.0 | 1.567 us | 0.0108 us | 0.0095 us | 0.1926 |   1.19 KB |
| ParseMessage |  CsProjnet461 | 1.363 us | 0.0230 us | 0.0204 us | 0.2003 |   1.24 KB |

On faster machines, the branch is :dash:&mdash;my local output (i7-8700K, 32 GB of RAM, W10 1709) shows 924ns and 799ns for .NET Core and .NET Desktop, respectively.

Fixes #3.
Fixes #5.
Fixes #6.
Fixes #14.